requirements.txt
streamlit>=1.32.0
pandas>=2.2.1
plotly>=5.20.0
google-generativeai>=0.4.1
python-dotenv>=1.0.1
